board: Add Libre Computer AML-S905X-CC board support
This patch adds support for the Libre Computer AML-S905X-CC based
on an Amlogic S905X SoCs.

The board features :
- 4xCortex-A53 Cores
- 1 or 2Gib or DDR4
- RS232 TTL Header
- MicroUSB for Power Only
- HDMI 2.0b output + CEC
- CVBS + Audio Jack
- 10/100 Ethernet
- 4xUSB2.0
- IR Receiver
- SDCard
- eMMC optional slot
- 40pin connector partially RPi compatible
- 8Pin I2S 7.1 connector
- 3Pin SPDIF connector

The board is entirely supported since U-boot 2018.01 and Linux 4.14.

The Buildroot support is done with upstream only package and the
previously submitted "aml-s905x-cc-fip" package to generate a valid
bootable SDCard image.